pyproject.toml
[project]
name = "skill-quiver"
version = "0.1.0"
description = "CLI tool for managing curated skill collections"

license = "MIT"
requires-python = ">=3.13"
dependencies = [
    "kdl-py>=1.1.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.scripts]
quiv = "skill_quiver.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skill_quiver"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "respx>=0.22",
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py313"

[tool.pytest.ini_options]
testpaths = ["tests"]
